requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cloudinary==1.25.0
confusable-homoglyphs==3.2.0
Django==3.2.2
django-bootstrap3==15.0.0
django-registration==3.1.2
djangorestframework==3.12.4
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
